+++ controllers/products.js
@@ -1,0 +1,26 @@
+const products = [];
+
+exports.getAddProduct = (req, res, next) => {
+  //res.sendFile(path.join(rootDir, 'views', 'add-product.html'));
+  res.render('add-product', {
+    pageTitle: 'Add Product',
+    path: '/admin/add-product'
+  });
+}
+
+exports.postAddProduct = (req, res, next) => {
+  products.push({title: req.body.title});
+
+  res.redirect('/');
+}
+
+exports.getProducts = (req, res, next) => {
+  res.render('shop', {pageTitle: 'Shop',
+                      prods: products,
+                      docTitle: 'Shop',
+                      path: '/',
+                      hasProducts: products.length > 0,
+                      activeShop: true,
+                      productCSS: true
+                    });
+}
