+++ routes/admin.js
@@ -2,22 +2,12 @@ const express = require('express');
 
 const path = require('path');
 
-const rootDir = require('../util/path');
-
 const router = express.Router();
 
-const products = [];
+const productsController = require('../controllers/products');
 
-router.get('/add-product',(req, res, next) => {
-  //res.sendFile(path.join(rootDir, 'views', 'add-product.html'));
-  res.render('add-product', {pageTitle: 'Add Product', path: '/admin/add-product'})
-});
+router.get('/add-product', productsController.getAddProduct);
 
-router.post('/add-product',(req, res, next) => {
-  products.push({title: req.body.title});
+router.post('/add-product', productsController.postAddProduct);
 
-  res.redirect('/');
-});
-
-exports.routes = router;
-exports.products = products;
+module.exports = router;
