+++ routes/shop.js
@@ -2,24 +2,11 @@ const path = require('path');
 
 const express = require('express');
 
-const rootDir = require('../util/path');
+const productsController = require('../controllers/products');
 
 const router = express.Router();
 
-const adminData = require('./admin')
 
-router.get('/',(req, res, next) => {
-  // console.log(adminData.products);
-  // res.sendFile(path.join(rootDir, 'views', 'shop.html'));
-  const products = adminData.products;
-  res.render('shop', {pageTitle: 'Shop',
-                      prods: products,
-                      docTitle: 'Shop',
-                      path: '/',
-                      hasProducts: products.length > 0,
-                      activeShop: true,
-                      productCSS: true
-                    });
-});
+router.get('/', productsController.getProducts);
 
 module.exports = router;
